+++ algo.py
@@ -8,7 +8,7 @@ from collections import deque
 from numpy import trapz
 from scipy.stats import linregress
 
-from data.alerts import AlertCodes
+from data.alert import AlertCodes
 from data.measurements import Measurements
 from data.configurations import Configurations
 from sample_storage import SamplesStorage
