+++ tests/unit/graphics/test_alert_history_screen.py
@@ -5,7 +5,8 @@ import freezegun
 import pytest
 from unittest.mock import MagicMock
 
-from data.alerts import Alert, AlertCodes, AlertsQueue
+from data.alert import Alert, AlertCodes
+from data.alerts_queue import AlertsQueue
 from data.events import Events
 from drivers.driver_factory import DriverFactory
 from graphics.alerts_history.history_screen import HistoryScreen
