+++ tests/unit/test_no_breath.py
@@ -1,7 +1,8 @@
 import pytest
 
 from algo import Sampler
-from data import alerts
+from data import alert
+from data.alert import AlertCodes
 from data.measurements import Measurements
 from data.events import Events
 from data.configurations import Configurations
@@ -68,7 +69,7 @@ def test_sinus_alerts_when_no_breath(events, measurements, config):
     assert len(events.alerts_queue) == 1
 
     alert = events.alerts_queue.queue.get()
-    assert alert == alerts.AlertCodes.NO_BREATH
+    assert alert == AlertCodes.NO_BREATH
 
 
 def test_dead_man_alerts_when_no_breath(events, measurements, config):
@@ -94,7 +95,7 @@ def test_dead_man_alerts_when_no_breath(events, measurements, config):
     assert len(events.alerts_queue) >= 1
 
     all_alerts = list(events.alerts_queue.queue.queue)
-    assert all(alert == alerts.AlertCodes.NO_BREATH for alert in all_alerts)
+    assert all(alert == AlertCodes.NO_BREATH for alert in all_alerts)
 
 
 def test_noise_alerts_when_no_breath(events, measurements, config):
@@ -120,4 +121,4 @@ def test_noise_alerts_when_no_breath(events, measurements, config):
     assert len(events.alerts_queue) >= 1
 
     all_alerts = list(events.alerts_queue.queue.queue)
-    assert all(alert == alerts.AlertCodes.NO_BREATH for alert in all_alerts)
+    assert all(alert == AlertCodes.NO_BREATH for alert in all_alerts)
