+++ tests/unit/test_volume.py
@@ -5,7 +5,8 @@ import pytest
 from pytest import approx
 
 from algo import Sampler
-from data import alerts
+from data import alert
+from data.alert import AlertCodes
 from data.measurements import Measurements
 from data.events import Events
 from data.configurations import Configurations
@@ -102,7 +103,7 @@ def test_sampler_alerts_when_volume_exceeds_minium(events, measurements, config,
     assert len(events.alerts_queue) > 0
 
     all_alerts = list(events.alerts_queue.queue.queue)
-    assert all(alert == alerts.AlertCodes.VOLUME_LOW for alert in all_alerts)
+    assert all(alert == AlertCodes.VOLUME_LOW for alert in all_alerts)
 
 
 def test_sampler_alerts_when_volume_exceeds_maximum(events, measurements, config, driver_factory):
@@ -126,4 +127,4 @@ def test_sampler_alerts_when_volume_exceeds_maximum(events, measurements, config
     assert len(events.alerts_queue) > 0
 
     all_alerts = list(events.alerts_queue.queue.queue)
-    assert all(alert == alerts.AlertCodes.VOLUME_HIGH for alert in all_alerts)
+    assert all(alert == AlertCodes.VOLUME_HIGH for alert in all_alerts)
